platform/requirements.txt
# Core AI Frameworks
semantic-kernel>=0.9.0
autogen-agentchat>=0.2.0
langchain>=0.1.0
langchain-community>=0.0.20
langgraph>=0.0.40

# AI Model Providers
openai>=1.10.0
anthropic>=0.8.0
azure-identity>=1.15.0

# Async & Performance
aiohttp>=3.9.0
asyncio>=3.4.3
tenacity>=8.2.3

# Utilities
python-dotenv>=1.0.0
pydantic>=2.5.0
pyyaml>=6.0.1
rich>=13.7.0

# Optional: Development Tools
pytest>=7.4.0
black>=23.12.0
mypy>=1.8.0
